% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeSSD.R
\name{BayeSSD}
\alias{BayeSSD}
\title{Title}
\usage{
BayeSSD(
  eta = 0.8,
  attrition = "weibull",
  params = c(0.5, 1),
  m = 100,
  t.points = c(0, 1, 2, 3, 4),
  var.u0 = 0.01,
  var.u1 = 0.1,
  var.e = 0.01,
  cov = 0,
  eff.sizes = c(0, 0.8, 0.8),
  BFthres = 5,
  fraction = 1,
  log.grow = F,
  seed = NULL,
  hypothesis = "a<b<c",
  PMPthres = 0.9,
  sensitivity = F,
  tol = 0.001,
  N_max = 1000
)
}
\arguments{
\item{eta}{The desired power level.
#'}

\item{attrition}{The attrition pattern (FALSE for no attrition, otherwise "weibull", "modified_weibull", "linear_exponential", "log_logistic", "gompertz" or "non-parametric")}

\item{params}{The parameters passed to the survival function specified in "attrition". First parameter is omega, second is gamma.}

\item{m}{The number of datasets simulated in each iteration. The higher m, the more accurate the power level but the higher the computation cost}

\item{t.points}{The points in time of measurement. Can be non-equidistant.}

\item{var.u0}{The intercept variance.}

\item{var.u1}{The slope variance.}

\item{var.e}{The residual variance.}

\item{cov}{The covariance between intercept variance and slope variance.}

\item{eff.sizes}{The effect sizes defined as the differences between the regression coefficients of interaction between time and condition.}

\item{BFthres}{The Threshold a Bayes Factor needs to exceed in order to be considered convincing evidence.}

\item{fraction}{The fraction of information used to construct the prior for the Bayes Factor.}

\item{log.grow}{Use log-linear growth?#' @param var.u0}

\item{seed}{Set a seed for reproducibility}

\item{hypothesis}{The hypothesis to be evaluated. Treatment groups are coded as "a", "b", "c", etc.}

\item{PMPthres}{The Threshold a Posterior Model Probability needs to exceed in order to be considered convincing evidence.}

\item{sensitivity}{Logical. Conduct a sensitivity analysis for the parameter fraction?}

\item{tol}{Tolerance for the deviation of the final result from eta. Higher values may speed up performance.}

\item{N_max}{The maximum sample size to be considered. Lower values speed up performance.}
}
\value{
Returns the sample size (number of subjects) necessary to achieve the desired power level (eta).
}
\description{
Title
}
