% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aafbf.R
\name{calc_aafbf}
\alias{calc_aafbf}
\title{CALCULATE AAFBF}
\usage{
calc_aafbf(type, estimates, sigma, b, n_eff)
}
\arguments{
\item{type}{String that indicates the type of comparison of hypotheses. "equality"}

\item{estimates}{R object with estimates.}

\item{sigma}{list with covariances.}

\item{b}{Numerical. Fraction of information to specify the prior.}

\item{n_eff}{Effective sample size.}

\item{n}{numeric. Effective sample size.}
}
\description{
aafbf
}
