% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hyp.R
\name{check_hyp}
\alias{check_hyp}
\title{Check if hypothesis is true according to 'eff.sizes'}
\usage{
check_hyp(hypothesis, eff.sizes)
}
\arguments{
\item{hypothesis}{The research hypothesis of interest, which must be true}

\item{eff.sizes}{The vector of true effect sizes}
}
\value{
TRUE if the hypothesis is true or FALSE if it is not true
}
\description{
Check if hypothesis is true according to 'eff.sizes'
}
\examples{
check_hyp("a<b<c", c(0, .5, .8)) # returns TRUE
}
