% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gen_CRT_data}
\alias{gen_CRT_data}
\title{Generation of data sets for two treatment-condition cluster randomized trial}
\usage{
gen_CRT_data(
  ndatasets = ndatasets,
  n1 = n1,
  n2 = n2,
  var_u0 = var_u0,
  var_e = var_e,
  mean_interv,
  batch_size
)
}
\arguments{
\item{ndatasets}{Numeric. Number of data sets that the user wants to generate to determine the sample size}

\item{n1}{Numeric. Cluster size}

\item{n2}{Numeric. Total number of clusters.}

\item{var_u0}{Numeric. Between-cluster variance. Variance at the cluster level.}

\item{var_e}{Numeric. Within-cluster variance.}

\item{mean_interv}{Numeric. Equivalent to the effect size.}

\item{batch_size}{This parameter determines the size of batches used during the fitting of the multilevel model.}

\item{rho}{Numeric. Intraclass correlation}
}
\description{
data generation
}
\examples{
gen_CRT_data(ndatasets=10, n1=30, n2=10, var_u0=.01, var_e=.1, mean_interv=.5, batch_size=5)
}
