% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neff.R
\name{get_neff}
\alias{get_neff}
\title{Calculate the effective sample size in a multilevel model with attrition for multiple treatment groups}
\usage{
get_neff(model, N, t.points, surviv)
}
\arguments{
\item{model}{A multilevel model of class 'mermod' (specifically 'lmerMod') with ≥2 treatment conditions}

\item{N}{Number of subjects in the model}

\item{t.points}{Vector containing the measurement occasions}

\item{surviv}{Vector (or list of vectors) containing the proportion of subjects remaining at each timepoint}
}
\value{
Returns the effective sample size for each treatment condition's effects
}
\description{
Calculate the effective sample size in a multilevel model with attrition for multiple treatment groups
}
\examples{
# Example with 3 treatment conditions
library(lme4)
library(Matrix)
library(MASS)
get_neff_mis_mv(model = lmermodel, N=100, t.points=c(0,1,2,3,4),
               surviv=list(c(1,.9,.8,.6,.5), c(1,.85,.7,.5,.3), c(1,.95,.9,.8,.7)))
}
