% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neff.R
\name{get_neff_mis_mv}
\alias{get_neff_mis_mv}
\title{Calculate the effective sample size in a multilevel model with attrition for three treatment groups}
\usage{
get_neff_mis_mv(model, N, t.points, surviv)
}
\arguments{
\item{model}{A multilevel model of class 'mermod' (more specifically, an object of subclass 'lmerMod')}

\item{N}{Number of subjects in the model.}

\item{t.points}{Vector containing the measurement occasions.}

\item{surviv}{Vector (or list of vectors) containing the proportion of subjects remaining still in the study at each timepoint as returned by the function "survival".}
}
\value{
Returns the effective sample size for each of the three treatment conditions.
}
\description{
Calculate the effective sample size in a multilevel model with attrition for three treatment groups
}
\examples{
library(lme4)
library(Matrix)
library(MASS)
get_neff_mis_mv(model = lmermodel, N=100, t.points=c(0,1,2,3,4), surviv=c(1, .9, .8, .6, .5))
}
