% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_results}
\alias{print_results}
\title{print results}
\usage{
print_results(object_result)
}
\arguments{
\item{object_result}{object containing the result}
}
\description{
print results
}
\details{
Print results for CRT
}
