% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results_longit.R
\name{print_results_SSD_longit}
\alias{print_results_SSD_longit}
\title{print results for SSD_longit}
\usage{
print_results_SSD_longit(result)
}
\arguments{
\item{result}{object containing the result of SSD_longit}
}
\description{
print results
}
\details{
Print results for the function SSD_longit
}
