% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD_clusters_inform.R
\name{SSD_crt_inform}
\alias{SSD_crt_inform}
\title{Sample Size Determination for Cluster Randomized Trials}
\usage{
SSD_crt_inform(
  eff_size,
  n1 = 15,
  n2 = 30,
  ndatasets = 1000,
  rho,
  BF_thresh,
  eta = 0.8,
  fixed = "n2",
  max = 1000,
  batch_size = 100
)
}
\arguments{
\item{n1}{Numeric. Cluster size}

\item{n2}{Numeric. Total number of clusters.}

\item{rho}{Numeric. Intraclass correlation}

\item{eta}{Numeric. Probability of exceeding Bayes Factor threshold.}

\item{fixed}{Character. Indicating which sample is fixed (n1 or n2)}

\item{max}{Maximum sample size.}

\item{batch_size}{This parameter determines the size of batches used during the fitting of the multilevel model.}

\item{eff.size}{Numeric. Effect size}

\item{n.datasets}{Numeric. Number of data sets that the user wants to generate to determine the sample size.}

\item{BF.thresh}{Numeric. Value of the Bayes factor that is going to be the threshold.}
}
\value{
Returns the result of the SSD procedure for CRT
}
\description{
SSD for CRT
}
\examples{
SSD_crt_inform(eff_size, n1 = 15, n2 = 30,
ndatasets = 1000, rho, BF_thresh, eta = 0.8,
fixed = 'n2', max = 1000, batch_size = 100)
}
