% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD_clusters_function.R
\name{SSD_crt_null}
\alias{SSD_crt_null}
\title{Sample Size Determination for Cluster Randomized Trials}
\usage{
SSD_crt_null(
  eff_size,
  n1 = 15,
  n2 = 30,
  ndatasets = 1000,
  rho,
  BF_thresh1,
  BF_thresh0,
  eta1 = 0.8,
  eta0 = 0.8,
  fixed = "n2",
  b_fract = 3,
  max = 1000,
  batch_size = 100
)
}
\arguments{
\item{eff_size}{Numeric. Effect size}

\item{n1}{Numeric. Cluster size}

\item{n2}{Numeric. Total number of clusters}

\item{ndatasets}{Numeric. Number of data sets that the user wants to generate to determine the sample size.}

\item{rho}{Numeric. Intraclass correlation}

\item{fixed}{Character. Indicating which sample is fixed (n1 or n2)}

\item{b_fract}{Numeric. Fraction of information used to specify the prior distribution.}

\item{max}{Maximum sample size.}

\item{batch_size}{This parameter determines the size of batches used during the fitting of the multilevel model.}

\item{BF_thresh}{Numeric. Value of the Bayes factor that is going to be the threshold.}

\item{eta}{Numeric. Probability of exceeding Bayes Factor threshold.}
}
\value{
Returns the result of the SSD procedure for CRT
}
\description{
SSD for CRT
}
\examples{
SSD_crt_null(eff_size, n1 = 15, n2 = 30, ndatasets = 1000, rho, BF_thresh1, BF_thresh0, eta1 = 0.8, eta0 = 0.8, fixed = "n2", b_fract = 3, max = 1000, batch_size = 100)
}
