% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeSSD.R
\name{SSD_longit}
\alias{SSD_longit}
\title{Perform simulation-based Bayesian sample size determination}
\usage{
SSD_longit(
  eta = 0.8,
  hypothesis = "a<b<c",
  eff.sizes = c(0, 0.5, 0.8),
  BFthres = 5,
  PMPthres = 0.9,
  method = "bfc",
  attrition = "weibull",
  params = c(0.5, 1),
  t.points = c(0, 1, 2, 3, 4),
  var.u0 = 0.01,
  var.u1 = 0.1,
  var.e = 0.01,
  cov = 0,
  m = 10000,
  log.grow = F,
  seed = NULL,
  sensitivity = F,
  tol = 0.01,
  N.max = 1000,
  N.min = 30
)
}
\arguments{
\item{eta}{The desired power level.}

\item{hypothesis}{List of the hypotheses to be evaluated. Treatment groups are coded as \code{a}, \code{b}, \code{c}, etc.}

\item{eff.sizes}{The effect sizes defined as the differences between the regression coefficients of interaction between time and condition.}

\item{BFthres}{The Threshold a Bayes Factor needs to exceed in order to be considered convincing evidence.}

\item{PMPthres}{The Threshold a Posterior Model Probability needs to exceed in order to be considered convincing evidence.}

\item{method}{The method used for hypothesis evaluation. If \code{bfc}/\code{BFc}, then the hypothesis is compared against its complement via the Bayes Factor. If \code{bf}/\code{BF}, then the first hypothesis is compared to the second one via the Bayes Factor. If \code{pmp}/\code{PMP}, then the first hypothesis is compared to the whole set of hypotheses including the complement via posterior model probabilities.}

\item{attrition}{The attrition pattern (\code{FALSE} for no attrition, otherwise \code{weibull}, \code{modified_weibull}, \code{linear_exponential}, \code{log_logistic}, \code{gompertz} or \code{non-parametric})}

\item{params}{The parameters passed to the survival function specified in \code{attrition}. First parameter is omega, second gamma and third (if applicable) is kappa.}

\item{t.points}{The points in time of measurement. Can be non-equidistant.}

\item{var.u0}{The intercept variance.}

\item{var.u1}{The slope variance.}

\item{var.e}{The residual variance.}

\item{cov}{The covariance between intercept variance and slope variance.}

\item{m}{The number of datasets simulated in each iteration. The higher \code{m}, the more accurate the power level but the higher the computation time}

\item{log.grow}{Logical. Use log-linear growth?}

\item{seed}{Set a seed for reproducibility?}

\item{sensitivity}{Logical. Conduct a sensitivity analysis for the b fraction?}

\item{tol}{Tolerance for the deviation of the final result from \code{eta}. Higher values may speed up performance.}

\item{N.max}{The maximum sample size to be considered. Lower values may speed up performance.}

\item{N.min}{The minimum sample size to be considered. Higher values may speed up performance.}
}
\value{
Returns the sample size (number of subjects) necessary to achieve the desired power level \code{eta}.
}
\description{
Perform simulation-based Bayesian sample size determination
}
\examples{
SSD_longit(eta=.8, attrition="weibull", params=c(.5,1),
m=100, t.points=c(0,1,2,3,4), var.u0=0.01,
var.u1=.1, var.e=.01, cov=0, eff.sizes=c(0, .8, .8),
BFthres=5, log.grow=F, seed=NULL,
hypothesis="a<b<c", PMPthres=.9, sensitivity=F, tol=.001,
N.max=1000)
}
