% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{Calculate survival curves to model attrition}
\usage{
survival(distributions, params, t.points)
}
\arguments{
\item{distributions}{Can be either a vector of length 1 or of length g, where g is the number of treatment conditions. Specifies the distribution(s) used to model attrition.}

\item{params}{List with either 1 element or g elements. Each element contains the parameters for the respective survival function.}

\item{t.points}{Vector containing the measurement occasions.}
}
\value{
Returns the proportion of individuals still remaining in the study at each timepoint.
}
\description{
Calculate survival curves to model attrition
}
\examples{
survival(distributions="weibull", params=list(c(.5, 1)), t.points=c(0,1,2,3,4))
}
