% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSSD-package.R
\docType{package}
\name{BayesSSD-package}
\alias{BayesSSD}
\alias{BayesSSD-package}
\title{BayesSSD: Bayesian sample size determination for longitudinal experiments}
\description{
This package performs simulation-based Bayesian sample size determination for longitudinal experiments.
}
\author{
\strong{Maintainer}: Ulrich Lösener \email{u.c.losener1@uu.nl}

}
\keyword{internal}
